restrict deletion of shared users unless their shared type is "INVITED".

+++ components/org.wso2.carbon.identity.organization.management.organization.user.sharing/src/main/java/org/wso2/carbon/identity/organization/management/organization/user/sharing/listener/SharedUserOperationEventListener.java
@@ -32,6 +32,7 @@ import org.wso2.carbon.identity.core.util.IdentityUtil;
 import org.wso2.carbon.identity.core.util.LambdaExceptionUtils;
 import org.wso2.carbon.identity.organization.management.organization.user.sharing.OrganizationUserSharingService;
 import org.wso2.carbon.identity.organization.management.organization.user.sharing.OrganizationUserSharingServiceImpl;
+import org.wso2.carbon.identity.organization.management.organization.user.sharing.constant.SharedType;
 import org.wso2.carbon.identity.organization.management.organization.user.sharing.internal.OrganizationUserSharingDataHolder;
 import org.wso2.carbon.identity.organization.management.organization.user.sharing.models.UserAssociation;
 import org.wso2.carbon.identity.organization.management.organization.user.sharing.util.OrganizationSharedUserUtil;
@@ -57,6 +58,7 @@ import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 
 import static org.wso2.carbon.identity.claim.metadata.mgt.util.ClaimConstants.SHARED_PROFILE_VALUE_RESOLVING_METHOD;
@@ -64,10 +66,10 @@ import static org.wso2.carbon.identity.claim.metadata.mgt.util.ClaimConstants.Sh
 import static org.wso2.carbon.identity.claim.metadata.mgt.util.ClaimConstants.SharedProfileValueResolvingMethod.FROM_ORIGIN;
 import static org.wso2.carbon.identity.claim.metadata.mgt.util.ClaimConstants.SharedProfileValueResolvingMethod.FROM_SHARED_PROFILE;
 import static org.wso2.carbon.identity.organization.management.organization.user.sharing.constant.UserSharingConstants.CLAIM_MANAGED_ORGANIZATION;
+import static org.wso2.carbon.identity.organization.management.organization.user.sharing.constant.UserSharingConstants.ErrorMessage.ERROR_CODE_UNAUTHORIZED_DELETION_OF_SHARED_USER;
 import static org.wso2.carbon.identity.organization.management.service.constant.OrganizationManagementConstants.ErrorMessages.ERROR_CODE_MANAGED_ORGANIZATION_CLAIM_UPDATE_NOT_ALLOWED;
 import static org.wso2.carbon.identity.organization.management.service.constant.OrganizationManagementConstants.ErrorMessages.ERROR_CODE_SHARED_USER_CLAIM_UPDATE_NOT_ALLOWED;
 import static org.wso2.carbon.identity.organization.management.service.util.Utils.getOrganizationId;
-import static org.wso2.carbon.identity.organization.management.service.util.Utils.getTenantDomain;
 
 /**
  * User operation event listener for shared user management.
@@ -97,15 +99,22 @@ public class SharedUserOperationEventListener extends AbstractIdentityUserOperat
             // The organization where the user identity is managed. Clear all the associations of the user.
             String associatedOrgId = OrganizationSharedUserUtil
                     .getUserManagedOrganizationClaim((AbstractUserStoreManager) userStoreManager, userID);
-            if (associatedOrgId != null) {
-                // User is associated only for shared users. Hence, delete the user association.
-                return organizationUserSharingService.deleteUserAssociation(userID, associatedOrgId);
-            }
-
             String orgId = getOrganizationId();
-            if (orgId == null) {
-                orgId = OrganizationUserSharingDataHolder.getInstance().getOrganizationManager()
-                        .resolveOrganizationId(getTenantDomain());
+
+            if (associatedOrgId != null) {
+                // Retrieve the user association details for the given user and organization.
+                UserAssociation userAssociation = getUserAssociation(userID, orgId);
+                String sharedType = userAssociation.getSharedType();
+
+                // Restrict deletion if the shared type is not "INVITED".
+                if (!Objects.equals(sharedType, SharedType.INVITED.name())) {
+                    throw new UserStoreClientException(
+                            ERROR_CODE_UNAUTHORIZED_DELETION_OF_SHARED_USER.getDescription(),
+                            ERROR_CODE_UNAUTHORIZED_DELETION_OF_SHARED_USER.getCode());
+                }
+
+                // Delete the user's association with the organization.
+                return organizationUserSharingService.deleteUserAssociation(userID, associatedOrgId);
             }
 
             // Delete all the user associations of the user.
